third_party/blink/web_tests/external/wpt/html/interaction/focus/processing-model/focusVisible-expected.txt

This is a testharness.js-based test.
[FAIL] FocusOptions.focusVisible: button
  assert_equals: :focus-visible expected true but got false
[FAIL] FocusOptions.focusVisible: input
  assert_equals: :focus-visible expected false but got true
[FAIL] FocusOptions.focusVisible: #contenteditable
  assert_equals: :focus-visible expected false but got true
[FAIL] FocusOptions.focusVisible: #tabindex
  assert_equals: :focus-visible expected true but got false
Harness: the test ran to completion.
